backend/requirements.txt
aiohttp==3.8.5 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
aiosignal==1.3.1 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
anyio==3.7.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
async-timeout==4.0.2 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
attrs==23.1.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
babel==2.12.1 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
bcrypt==4.0.1 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
certifi==2023.5.7 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
charset-normalizer==3.1.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
click==8.1.3 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
colorama==0.4.6 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0" and platform_system == "Windows"
deta[async]==1.2.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
docxcompose==1.4.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
docxtpl==0.16.7 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
ecdsa==0.18.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
exceptiongroup==1.1.1 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
fastapi==0.99.1 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
frozenlist==1.4.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
h11==0.14.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
idna==3.4 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
jinja2==3.1.2 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
lxml==4.9.3 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
markupsafe==2.1.3 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
multidict==6.0.4 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
passlib[brypt]==1.7.4 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
pyasn1==0.5.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
pydantic==1.10.10 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
python-docx==0.8.11 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
python-jose==3.3.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
python-multipart==0.0.6 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
requests==2.31.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
rsa==4.9 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
setuptools==68.0.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
six==1.16.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
sniffio==1.3.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
starlette==0.27.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
types-passlib==1.7.7.12 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
types-pyasn1==0.4.0.5 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
types-python-jose==3.3.4.7 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
typing-extensions==4.7.1 ; python_full_version >= "3.9.0" and python_version < "3.10"
urllib3==2.0.3 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
uvicorn==0.22.0 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
yarl==1.9.2 ; python_full_version >= "3.9.0" and python_full_version < "3.10.0"
